+++ packages/cli.cmd.typescript/.eslintrc.cjs
