+++ packages/cli.cmd.typescript/bin/osdk.mjs
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+// @ts-check
+
+import { cli } from "../build/js/index.mjs";
+cli();
