+++ packages/cli.cmd.typescript/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@osdk/cli.unstable.typescript.generate",
+  "name": "@osdk/cli.cmd.typescript",
   "version": "0.0.0",
   "license": "Apache-2.0",
   "repository": {
@@ -8,12 +8,12 @@
   },
   "exports": {
     ".": {
-      "types": "./build/types/src/index.d.ts",
+      "types": "./build/types/index.d.ts",
       "import": "./build/js/index.mjs",
       "require": "./build/js/index.cjs"
     },
     "./*": {
-      "types": "./build/types/src/public/*.d.ts",
+      "types": "./build/types/public/*.d.ts",
       "import": "./build/js/public/*.mjs",
       "require": "./build/js/public/*.cjs"
     }
@@ -26,7 +26,20 @@
     "transpile": "tsup",
     "typecheck": "tsc-absolute --build"
   },
+  "dependencies": {
+    "@arethetypeswrong/cli": "^0.15.2",
+    "@osdk/cli.common": "workspace:*",
+    "@osdk/gateway": "workspace:*",
+    "@osdk/generator": "workspace:*",
+    "@osdk/shared.net": "workspace:*",
+    "consola": "^3.2.3",
+    "find-up": "^7.0.0",
+    "tslib": "^2.6.2",
+    "yargs": "^17.7.2"
+  },
   "devDependencies": {
+    "@types/node": "^18.0.0",
+    "@types/yargs": "^17.0.29",
     "typescript": "^5.4.2"
   },
   "publishConfig": {
@@ -43,5 +56,6 @@
   ],
   "main": "./build/js/index.cjs",
   "module": "./build/js/index.mjs",
-  "types": "./build/types/src/index.d.ts"
+  "types": "./build/types/index.d.ts",
+  "type": "module"
 }
