+++ packages/cli.common/src/commands/auth/CommonAuthArgs.ts
@@ -14,10 +14,8 @@
  * limitations under the License.
  */
 
-export type { CliCommonArgs } from "./CliCommonArgs.js";
-export type { CommonAuthArgs } from "./commands/auth/CommonAuthArgs.js";
-export type { LoginArgs } from "./commands/auth/login/LoginArgs.js";
-export { ExitProcessError } from "./ExitProcessError.js";
-export { getYargsBase } from "./getYargsBase.js";
-export { isValidSemver } from "./util/isValidSemver.js";
-export { YargsCheckError } from "./YargsCheckError.js";
+import type { CliCommonArgs } from "../../CliCommonArgs.js";
+
+export interface CommonAuthArgs extends CliCommonArgs {
+  foundryUrl: string;
+}
