+++ packages/cli.common/src/commands/auth/login/token.ts
