--- packages/cli.unstable.typescript.generate/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../monorepo/tsconfig/tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build/types",
-    "composite": true
-  },
-  "include": [
-    "./src/**/*",
-    ".eslintrc.cjs",
-    "package.json"
-  ],
-  "references": []
-}
