+++ packages/cli/src/commands/site/index.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
+import type { CliCommonArgs } from "@osdk/cli.common";
+import { YargsCheckError } from "@osdk/cli.common";
 import type { CommandModule } from "yargs";
-import type { CliCommonArgs } from "../../CliCommonArgs.js";
 import type { ThirdPartyAppRid } from "../../net/ThirdPartyAppRid.js";
 import type { LoadedFoundryConfig } from "../../util/config.js";
 import configLoader from "../../util/configLoader.js";
-import { YargsCheckError } from "../../YargsCheckError.js";
 import type { CommonSiteArgs } from "./CommonSiteArgs.js";
 import deploy from "./deploy/index.js";
 import { logSiteCommandConfigFileOverride } from "./logSiteCommandConfigFileOverride.js";
