+++ packages/cli/tsconfig.json
@@ -14,6 +14,12 @@
       "path": "../api"
     },
     {
+      "path": "../cli.cmd.typescript"
+    },
+    {
+      "path": "../cli.common"
+    },
+    {
       "path": "../gateway"
     },
     {
